move transport related test helpers to transport_test.go
